open-campus-demo-v1/services/visitor_tracker/requirements.txt
ultralytics
opencv-python
numpy
langchain
langchain-openai
python-dotenv
